r/aws_efs_replication_configuration: Tidy up.
