+++ internal/service/efs/find.go
@@ -62,32 +62,3 @@ func FindFileSystemPolicyByID(ctx context.Context, conn *efs.EFS, id string) (*e
 
 	return output, nil
 }
-
-func FindReplicationConfigurationByID(ctx context.Context, conn *efs.EFS, id string) (*efs.ReplicationConfigurationDescription, error) {
-	input := &efs.DescribeReplicationConfigurationsInput{
-		FileSystemId: aws.String(id),
-	}
-
-	output, err := conn.DescribeReplicationConfigurationsWithContext(ctx, input)
-
-	if tfawserr.ErrCodeEquals(err, efs.ErrCodeFileSystemNotFound) || tfawserr.ErrCodeEquals(err, efs.ErrCodeReplicationNotFound) {
-		return nil, &retry.NotFoundError{
-			LastError:   err,
-			LastRequest: input,
-		}
-	}
-
-	if err != nil {
-		return nil, err
-	}
-
-	if output == nil ||
-		len(output.Replications) == 0 ||
-		output.Replications[0] == nil ||
-		len(output.Replications[0].Destinations) == 0 ||
-		output.Replications[0].Destinations[0] == nil {
-		return nil, tfresource.NewEmptyResultError(input)
-	}
-
-	return output.Replications[0], nil
-}
